Fix cut insertion and SCLookUp
FindOne cut insertion fails because SCIterDel reorders the variables being scanned, emeaning the cut will be on the wrong variable. Fix by calling SCIterDel before FindOneCuts.

SCLookUpIntro looks for SCIter followed directly by a Cut and replaces it with a SCLookUp. However, SCLookUp cannot be accomplished by a simple lookup anyway because it is possible that the lookup fails if the object is not of the correct subclass. Fix by just removing the concept of an SCLookUp.